Port src/marks/line from JS to TS (stage 1)
Combine line.js + line.d.ts into a single line.ts with inlined types.
No behaviour changes — runtime preserved verbatim from the .js file,
and types/JSDoc lifted from the .d.ts.
